initial submission - payment data processing tools
DRAFT version: this is work in progress and the code may be not running smoothly. The code can be used to process and analyze BACS payment data as produced and published by the ONS. The scripts were used to process confidential and experimental versions of the data, but may be also useful for analyzing current and future publicly available versions of the data.